Backend/requirements.txt
radon
semgrep
pyup
reportlab
sqlite3
